More restructuring, added setup file
