+++ .travis.yml
@@ -1,1 +1,9 @@
-language:python
+language: python
+python:
+  - "3.x"
+  - "3.4"
+before_install:
+  python setup.py install
+
+script:
+  python -m unittest app/test.py
