+++ app/api.py
@@ -1,7 +1,6 @@
 from flask import Flask, jsonify, request
-from .resources import resources
-APP = Flask(__name__)
-APP.config['JSONIFY_PRETTYPRINT_REGULAR'] = False
+from app import APP
+from app.resources import orders
 
 
 @APP.route('/orders/<int:order_id>', methods=['PUT'])
@@ -11,12 +10,12 @@ def update_status(order_id):
     '''
     data = request.get_json()
     status = data.get('status')
-    if not any(i.get('order_id') == order_id for i in resources.orders):
+    if not any(i.get('order_id') == order_id for i in orders):
         return 'Order does not exist', 404
-    for i in resources.orders:
+    for i in orders:
         if i.get('order_id') == order_id:
             i['order_status'] = status
-    return jsonify(resources.orders)
+    return jsonify(orders)
 
 
 if __name__ == '__main__':
