+++ run.py
@@ -1,4 +1,4 @@
-from app.api import APP
+from app import APP
 
 if __name__ == '__main__':
     APP.run(debug=True)
