+++ setup.py
@@ -1,0 +1,9 @@
+from setuptools import setup
+setup(
+    name='app',
+    packages=['app'],
+    include_package_data=True,
+    install_requires=[
+        'flask',
+    ],
+)
